Integrating waste classification
