+++ source/templates/classification.html
@@ -22,7 +22,7 @@
               <li class="nav-item active">
                 <a class="nav-link" href="classification.html">classification</a>
               </li>
-              <li class="nav-item active">
+              <li class="nav-item">
                 <a class="nav-link" href="learn-more.html">learn more</a>
               </li>
               <li class="nav-item">
@@ -33,7 +33,14 @@
       </nav>
     </div>
 
-    <h2 class="text-center">Tap to learn more about each kind of waste!</h2>
+    <h2 class="text-center">[Waste Category]</h2>
 
+    <img src="{{ url_for('video_feed_class') }}" class="rounded mx-auto d-block" alt="livestream">
+    <!-- <img src="http://192.168.1.50:8000/stream.mjpg">  -->
+  
+    <footer>
+      
+    </footer>
+  
   </body>
 </html>
