+++ source/templates/index.html
@@ -19,6 +19,9 @@
               <li class="nav-item">
                 <a class="nav-link" href="scan-page.html">scan</a>
               </li>
+              <li class="nav-item active">
+                <a class="nav-link" href="classification.html">classification</a>
+              </li>
               <li class="nav-item">
                 <a class="nav-link" href="learn-more.html">learn more</a>
               </li>
